requirements.txt
django<4
django-readers
django-debug-toolbar
djangorestframework
factory_boy
